Move `common` test into a single file
