--- tests/test_common/test_state.py
@@ -1,23 +1,0 @@
-from datetime import datetime
-
-from devtools import debug
-
-from fractal_server.common.schemas import _StateBase
-from fractal_server.common.schemas import StateRead
-
-
-def test_state():
-    s = _StateBase(data={"some": "thing"}, timestamp=datetime.now())
-    debug(s)
-    debug(s.sanitised_dict())
-    assert isinstance(s.sanitised_dict()["timestamp"], str)
-
-
-def test_state_read():
-    s = StateRead(data={"some": "thing"}, timestamp=datetime.now())
-    debug(s)
-    assert s.id is None
-
-    s = StateRead(data={"some": "thing"}, timestamp=datetime.now(), id=1)
-    debug(s)
-    assert s.id == 1
